Added adapter , decorator

+++ facade/src/main/java/pl/sda/mvc/model/Model.java
@@ -8,13 +8,13 @@ import java.util.List;
  */
 public class Model {
 
-    public static List<Integer> values=new ArrayList<>();
+    public static List<Long> values = new ArrayList<>();
 
-    public static List<Integer> getValues() {
+    public static List<Long> getValues() {
         return values;
     }
 
-    public static void addValues(int newValue){
+    public static void addValues(long newValue) {
         values.add(newValue);
     }
